backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pybaseball
pandas
scikit-learn
alembic 
python-dotenv
psycopg2 